[refactor] Dealer / Player 공통 로직 Participant로 분리
