+++ src/main/java/model/participant/Participant.java
@@ -1,6 +1,28 @@
 package model.participant;
 
+import model.card.Card;
+import model.card.Cards;
+
 public abstract class Participant {
-    // 카드를 뽑는다.
-    // 카드의 합을 계산한다.
+    protected final Cards cards;
+
+    protected Participant(Cards cards) {
+        this.cards = cards;
+    }
+
+    public Cards getCards() {
+        return cards;
+    }
+
+    public int getTotalValue() {
+        return cards.calculateScore();
+    }
+
+    public boolean isBust() {
+        return cards.calculateScore() >= bustThreshold();
+    }
+
+    protected abstract int bustThreshold();
+
+    public abstract Participant receive(Card card);
 }
